requirements.txt
python-telegram-bot[job-queue]>=20.0
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
python-dotenv
